+++ src/bookmarkManager/SideNav.js
@@ -1,6 +1,5 @@
 import React from "react";
-import Drawer from "@material-ui/core/Drawer";
-import Hidden from "@material-ui/core/Hidden";
+import { Drawer, Hidden } from "@material-ui/core";
 import SideNavContent from "./SideNavContent";
 
 export default function SideNav({
