+++ src/bookmarkManager/SideNavContent.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
-import BookmarkTreeView from "shared/components/bookmarks/BookmarkTreeView";
+import { BookmarkTreeView } from "shared/components/bookmarks";
 
 const useStyles = makeStyles(theme => ({
   toolbar: theme.mixins.toolbar,
