+++ src/shared/components/bookmarks/form/index.js
@@ -1,0 +1,1 @@
+export { default as BookmarkForm } from "./BookmarkForm";
