+++ src/shared/components/helpers/SearchInput.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import InputBase from "@material-ui/core/InputBase";
 import { fade, makeStyles } from "@material-ui/core/styles";
 import SearchIcon from "@material-ui/icons/Search";
-import useDebounce from "shared/lib/hooks";
+import { useDebounce } from "shared/lib/hooks";
 
 const useStyles = makeStyles(theme => ({
   search: {
