+++ src/shared/lib/hooks/index.js
@@ -1,1 +1,1 @@
-export { default } from "./useDebounce";
+export { default as useDebounce } from "./useDebounce";
